UVNX-216 주소 2줄 제한, 높이 계산
