+++ breadgood_app/lib/modules/register_bakery/model/duplicated_bakery.dart
@@ -1,0 +1,21 @@
+import 'dart:core';
+
+class CheckDuplicateBakery {
+  final int bakeryId;
+  final bool idDuplicate;
+  final String nickName;
+
+  CheckDuplicateBakery({
+    this.bakeryId,
+    this.idDuplicate,
+    this.nickName,
+  });
+
+  factory CheckDuplicateBakery.fromJson(Map<String, dynamic> json) {
+    return CheckDuplicateBakery(
+      bakeryId: json['bakeryId'],
+      idDuplicate: json['idDuplicate'],
+      nickName: json['nickName'],
+    );
+  }
+}
