+++ breadgood_app/lib/modules/register_bakery/screens/no_result.dart
@@ -1,0 +1,9 @@
+import 'dart:core';
+import 'package:flutter/material.dart';
+
+Padding getNoResult() => Padding(
+    padding: EdgeInsets.only(top: 107),
+    child: Align(
+      alignment: Alignment.center,
+      child: Image.asset('asset/images/cat_black_and_white.png'),
+    ));
